Reimplement task creation/termination chain logs for the webui
- Refactor and split out the "termui" and "webui" modules
