+++ aiomonitor/monitor.py
@@ -1,19 +1,17 @@
 from __future__ import annotations
 
 import asyncio
+import contextlib
 import contextvars
 import functools
 import logging
-import os
-import shlex
-import signal
 import sys
+import textwrap
 import threading
 import time
 import traceback
 import weakref
 from asyncio.coroutines import _format_coroutine  # type: ignore
-from contextvars import ContextVar, copy_context
 from datetime import timedelta
 from types import TracebackType
 from typing import (
@@ -23,33 +21,22 @@ from typing import (
     Dict,
     Final,
     Generator,
-    Iterable,
     List,
     Optional,
     Sequence,
-    TextIO,
-    Tuple,
+    Set,
     Type,
     TypeVar,
     cast,
 )
 
-import click
 import janus
 from aiohttp import web
-from click.parser import split_arg_string
-from click.shell_completion import CompletionItem, _resolve_context, _resolve_incomplete
-from prompt_toolkit import PromptSession
-from prompt_toolkit.application.current import get_app_session
-from prompt_toolkit.completion import CompleteEvent, Completer, Completion
-from prompt_toolkit.contrib.telnet.server import TelnetConnection, TelnetServer
-from prompt_toolkit.document import Document
-from prompt_toolkit.formatted_text import FormattedText
-from prompt_toolkit.shortcuts import print_formatted_text
-from terminaltables import AsciiTable
+from prompt_toolkit.contrib.telnet.server import TelnetServer
 
-from . import console
+from .exceptions import MissingTask
 from .task import TracedTask, persistent_coro
+from .termui.commands import interact
 from .types import (
     CancellationChain,
     FormattedLiveTaskInfo,
@@ -57,7 +44,6 @@ from .types import (
     TerminatedTaskInfo,
 )
 from .utils import (
-    AliasGroupMixin,
     _extract_stack_from_exception,
     _extract_stack_from_frame,
     _extract_stack_from_task,
@@ -66,26 +52,19 @@ from .utils import (
     _format_task,
     _format_terminated_task,
     _format_timedelta,
-    all_tasks,
-    cancel_task,
     get_default_args,
-    task_by_id,
 )
 from .webui.app import init_webui
 
 __all__ = (
     "Monitor",
-    "monitor_cli",
     "start_monitor",
 )
 
 log = logging.getLogger(__name__)
-current_monitor: ContextVar[Monitor] = ContextVar("current_monitor")
-current_stdout: ContextVar[TextIO] = ContextVar("current_stdout")
-command_done: ContextVar[asyncio.Event] = ContextVar("command_done")
 
 MONITOR_HOST: Final = "127.0.0.1"
-MONITOR_PORT: Final = 50101
+MONITOR_TERMUI_PORT: Final = 50101
 CONSOLE_PORT: Final = 50102
 MONITOR_WEBUI_PORT: Final = 50201
 
@@ -93,108 +72,29 @@ T = TypeVar("T")
 T_co = TypeVar("T_co", covariant=True)
 
 
-@click.group(cls=AliasGroupMixin, add_help_option=False)
-def monitor_cli():
-    """
-    To see the usage of each command, run them with "--help" option.
-    """
-    pass
+def task_by_id(
+    taskid: int, loop: asyncio.AbstractEventLoop
+) -> "Optional[asyncio.Task[Any]]":
+    tasks = all_tasks(loop=loop)
+    return next(filter(lambda t: id(t) == taskid, tasks), None)
 
 
-def _get_current_stdout() -> TextIO:
-    stdout = current_stdout.get(None)
-    if stdout is None:
-        return sys.stdout
+async def cancel_task(task: "asyncio.Task[Any]") -> None:
+    with contextlib.suppress(asyncio.CancelledError):
+        task.cancel()
+        await task
+
+
+def all_tasks(loop: asyncio.AbstractEventLoop) -> "Set[asyncio.Task[Any]]":
+    if sys.version_info >= (3, 7):
+        tasks = asyncio.all_tasks(loop=loop)
     else:
-        return stdout
-
-
-def _get_current_stderr() -> TextIO:
-    stdout = current_stdout.get(None)
-    if stdout is None:
-        return sys.stderr
-    else:
-        return stdout
-
-
-class ClickCompleter(Completer):
-    def __init__(self, root_command: click.BaseCommand) -> None:
-        self._root_command = root_command
-
-    def get_completions(
-        self,
-        document: Document,
-        complete_event: CompleteEvent,
-    ) -> Iterable[Completion]:
-        args = split_arg_string(document.current_line)
-        incomplete = document.get_word_under_cursor()
-        if incomplete and args and args[-1] == incomplete:
-            args.pop()
-        click_ctx = _resolve_context(self._root_command, {}, "", args)
-        cmd_or_param, incomplete = _resolve_incomplete(click_ctx, args, incomplete)
-        completions: List[CompletionItem] = cmd_or_param.shell_complete(
-            click_ctx, incomplete
-        )
-        start_position = document.find_backwards(incomplete) or 0
-        for completion in completions:
-            yield Completion(
-                completion.value,
-                start_position=start_position,
-            )
-
-
-def complete_task_id(
-    ctx: click.Context,
-    param: click.Parameter,
-    incomplete: str,
-) -> Iterable[str]:
-    # ctx here is created in the completer and does not have ctx.obj set as monitor.
-    # We take the monitor instance from the global context variable instead.
-    try:
-        self: Monitor = current_monitor.get()
-    except LookupError:
-        return []
-    return [
-        task_id
-        for task_id in map(str, sorted(map(id, all_tasks(loop=self._monitored_loop))))
-        if task_id.startswith(incomplete)
-    ][:10]
-
-
-def complete_trace_id(
-    ctx: click.Context,
-    param: click.Parameter,
-    incomplete: str,
-) -> Iterable[str]:
-    try:
-        self: Monitor = current_monitor.get()
-    except LookupError:
-        return []
-    return [
-        trace_id
-        for trace_id in map(str, sorted(self._terminated_tasks.keys()))
-        if trace_id.startswith(incomplete)
-    ][:10]
-
-
-def complete_signal_names(
-    ctx: click.Context,
-    param: click.Parameter,
-    incomplete: str,
-) -> Iterable[str]:
-    return [sig.name for sig in signal.Signals if sig.name.startswith(incomplete)]
+        tasks = asyncio.Task.all_tasks(loop=loop)
+    return tasks
 
 
 class Monitor:
     _event_loop_thread_id: Optional[int] = None
-    _cmd_prefix = "do_"
-    _empty_result = object()
-
-    intro = "\nAsyncio Monitor: {tasknum} task{s} running\nType help for available commands\n"  # noqa
-    help_template = "{cmd_name} {arg_list}\n    {doc}\n"
-    help_short_template = (
-        "    {cmd_name}{cmd_arg_sep}{arg_list}: {doc_firstline}"  # noqa
-    )
 
     console_locals: Dict[str, Any]
     _console_tasks: weakref.WeakSet[asyncio.Task[Any]]
@@ -218,7 +118,7 @@ class Monitor:
         loop: asyncio.AbstractEventLoop,
         *,
         host: str = MONITOR_HOST,
-        port: int = MONITOR_PORT,
+        termui_port: int = MONITOR_TERMUI_PORT,
         webui_port: int = MONITOR_WEBUI_PORT,
         console_port: int = CONSOLE_PORT,
         console_enabled: bool = True,
@@ -228,7 +128,7 @@ class Monitor:
     ) -> None:
         self._monitored_loop = loop or asyncio.get_running_loop()
         self._host = host
-        self._port = port
+        self._termui_port = termui_port
         self._webui_port = webui_port
         self._console_port = console_port
         self._console_enabled = console_enabled
@@ -239,10 +139,10 @@ class Monitor:
 
         self.prompt = "monitor >>> "
         log.info(
-            "Starting aiomonitor at telnet://%(host)s:%(port)d and http://%(host)s:%(wport)d",
+            "Starting aiomonitor at telnet://%(host)s:%(tport)d and http://%(host)s:%(wport)d",
             {
                 "host": host,
-                "port": port,
+                "tport": termui_port,
                 "wport": webui_port,
             },
         )
@@ -269,12 +169,12 @@ class Monitor:
 
     @property
     def port(self) -> int:
-        return self._port
+        return self._termui_port
 
     def __repr__(self) -> str:
         name = self.__class__.__name__
         return "<{name}: {host}:{port}>".format(
-            name=name, host=self._host, port=self._port
+            name=name, host=self._host, port=self._termui_port
         )
 
     def start(self) -> None:
@@ -287,11 +187,6 @@ class Monitor:
         self._event_loop_thread_id = threading.get_ident()
         self._ui_thread.start()
         self._ui_started.wait()
-
-        # Override the Click's stdout/stderr reference cache functions
-        # to let them use the correct stdout handler.
-        click.utils._default_text_stdout = _get_current_stdout
-        click.utils._default_text_stderr = _get_current_stderr
 
     @property
     def closed(self) -> bool:
@@ -323,7 +218,7 @@ class Monitor:
             self._monitored_loop.set_task_factory(self._original_task_factory)
             self._closed = True
 
-    def format_live_task_list(
+    def format_running_task_list(
         self, filter_: str, persistent: bool
     ) -> Sequence[FormattedLiveTaskInfo]:
         all_running_tasks = all_tasks(loop=self._monitored_loop)
@@ -423,6 +318,170 @@ class Monitor:
         else:
             raise ValueError("Invalid or non-existent task ID", task_id)
 
+    def format_running_task_stack(self, task_id: str | int):
+        depth = 0
+        task_id_ = int(task_id)
+        task = task_by_id(task_id_, self._monitored_loop)
+        if task is None:
+            raise MissingTask(task_id_)
+        task_chain: List[asyncio.Task[Any]] = []
+        while task is not None:
+            task_chain.append(task)
+            task_ref = self._created_traceback_chains.get(task)
+            task = task_ref() if task_ref is not None else None
+        prev_task = None
+        formatted_stack_list = []
+        for task in reversed(task_chain):
+            if depth == 0:
+                formatted_stack_list.append(
+                    (
+                        "header",
+                        (
+                            "Stack of the root task or coroutine scheduled "
+                            "in the event loop (most recent call last)"
+                        ),
+                    )
+                )
+            elif depth > 0:
+                assert prev_task is not None
+                formatted_stack_list.append(
+                    (
+                        "header",
+                        (
+                            "Stack of %s when creating the next task "
+                            "(most recent call last)" % _format_task(prev_task)
+                        ),
+                    )
+                )
+            stack = self._created_tracebacks.get(task)
+            if stack is None:
+                formatted_stack_list.append(
+                    (
+                        "content",
+                        (
+                            "No stack available (maybe it is a native code, "
+                            "a synchronous callback function, "
+                            "or the event loop itself)"
+                        ),
+                    )
+                )
+            else:
+                stack = _filter_stack(stack)
+                formatted_stack_list.append(
+                    (
+                        "content",
+                        textwrap.dedent("".join(traceback.format_list(stack))),
+                    )
+                )
+            prev_task = task
+            depth += 1
+        task = task_chain[0]
+        formatted_stack_list.append(
+            (
+                "header",
+                "Stack of %s (most recent call last):\n\n" % _format_task(task),
+            )
+        )
+        stack = _extract_stack_from_task(task)
+        if not stack:
+            formatted_stack_list.append(
+                (
+                    "content",
+                    "No stack available for %s" % _format_task(task),
+                )
+            )
+        else:
+            formatted_stack_list.append(
+                (
+                    "content",
+                    textwrap.dedent("".join(traceback.format_list(stack))),
+                )
+            )
+        return formatted_stack_list
+
+    def format_terminated_task_stack(self, trace_id: str):
+        depth = 0
+        tinfo_chain: List[TerminatedTaskInfo] = []
+        while trace_id is not None:
+            tinfo_chain.append(self._terminated_tasks[trace_id])
+            trace_id = self._canceller_chain.get(trace_id)  # type: ignore
+        prev_tinfo = None
+        formatted_stack_list = []
+        for tinfo in reversed(tinfo_chain):
+            if depth == 0:
+                formatted_stack_list.append(
+                    (
+                        "header",
+                        (
+                            "Stack of the root task or coroutine "
+                            "scheduled in the event loop"
+                            "(most recent call last)"
+                        ),
+                    )
+                )
+            elif depth > 0:
+                assert prev_tinfo is not None
+                formatted_stack_list.append(
+                    (
+                        "header",
+                        (
+                            "Stack of %s when creating the next task "
+                            "(most recent call last)"
+                            % _format_terminated_task(prev_tinfo)
+                        ),
+                    )
+                )
+            stack = tinfo.canceller_stack
+            if stack is None:
+                formatted_stack_list.append(
+                    (
+                        "content",
+                        (
+                            "No stack available "
+                            "(maybe it is a self-raised cancellation or exception)"
+                        ),
+                    )
+                )
+            else:
+                stack = _filter_stack(stack)
+                formatted_stack_list.append(
+                    (
+                        "content",
+                        textwrap.dedent("".join(traceback.format_list(stack))),
+                    )
+                )
+            prev_tinfo = tinfo
+            depth += 1
+        tinfo = tinfo_chain[0]
+        formatted_stack_list.append(
+            (
+                "header",
+                (
+                    "Stack of %s (most recent call last)"
+                    % _format_terminated_task(tinfo)
+                ),
+            )
+        )
+        stack = tinfo.termination_stack
+        if not stack:
+            formatted_stack_list.append(
+                (
+                    "content",
+                    (
+                        "No stack available for %s (the task has run to completion)"
+                        % _format_terminated_task(tinfo)
+                    ),
+                )
+            )
+        else:
+            formatted_stack_list.append(
+                (
+                    "content",
+                    textwrap.dedent("".join(traceback.format_list(stack))),
+                )
+            )
+        return formatted_stack_list
+
     async def _coro_wrapper(self, coro: Awaitable[T_co]) -> T_co:
         myself = asyncio.current_task()
         assert isinstance(myself, TracedTask)
@@ -479,7 +538,9 @@ class Monitor:
             self._ui_handle_cancellation_updates()
         )
         telnet_server = TelnetServer(
-            interact=self._interact, host=self._host, port=self._port
+            interact=functools.partial(interact, self),
+            host=self._host,
+            port=self._termui_port,
         )
         webui_app = await init_webui(self)
         webui_runner = web.AppRunner(webui_app)
@@ -541,452 +602,15 @@ class Monitor:
             self._canceller_stacks[update.target_id] = update.canceller_stack
             self._canceller_chain[update.target_id] = update.canceller_id
 
-    def print_ok(self, msg: str) -> None:
-        print_formatted_text(
-            FormattedText(
-                [
-                    ("ansibrightgreen", "✓ "),
-                    ("", msg),
-                ]
-            )
-        )
-
-    def print_fail(self, msg: str) -> None:
-        print_formatted_text(
-            FormattedText(
-                [
-                    ("ansibrightred", "✗ "),
-                    ("", msg),
-                ]
-            )
-        )
-
-    async def _interact(self, connection: TelnetConnection) -> None:
-        """
-        The interactive loop for each telnet client connection.
-        """
-        await asyncio.sleep(0.3)  # wait until telnet negotiation is done
-        tasknum = len(all_tasks(loop=self._monitored_loop))
-        s = "" if tasknum == 1 else "s"
-        print(self.intro.format(tasknum=tasknum, s=s), file=connection.stdout)
-        current_monitor_token = current_monitor.set(self)
-        current_stdout_token = current_stdout.set(connection.stdout)
-        # NOTE: prompt_toolkit's all internal console output automatically uses
-        #       an internal contextvar to keep the stdout consistent with the
-        #       current telnet connection.
-        prompt_session: PromptSession[str] = PromptSession(
-            completer=ClickCompleter(monitor_cli),
-            complete_while_typing=False,
-        )
-        lastcmd = "noop"
-        style_prompt = "#5fd7ff bold"
-        try:
-            while True:
-                try:
-                    user_input = (
-                        await prompt_session.prompt_async(
-                            FormattedText(
-                                [
-                                    (style_prompt, self.prompt),
-                                ]
-                            )
-                        )
-                    ).strip()
-                except KeyboardInterrupt:
-                    self.print_fail("To terminate, press Ctrl+D or type 'exit'.")
-                except (EOFError, asyncio.CancelledError):
-                    return
-                except Exception:
-                    self.print_fail(traceback.format_exc())
-                else:
-                    command_done_event = asyncio.Event()
-                    command_done_token = command_done.set(command_done_event)
-                    try:
-                        if not user_input and lastcmd is not None:
-                            user_input = lastcmd
-                        args = shlex.split(user_input)
-                        term_size = prompt_session.output.get_size()
-                        ctx = copy_context()
-                        ctx.run(
-                            monitor_cli.main,
-                            args,
-                            prog_name="",
-                            obj=self,
-                            standalone_mode=False,  # type: ignore
-                            max_content_width=term_size.columns,
-                        )
-                        await command_done_event.wait()
-                        if args[0] == "console":
-                            lastcmd = "noop"
-                        else:
-                            lastcmd = user_input
-                    except (click.BadParameter, click.UsageError) as e:
-                        self.print_fail(str(e))
-                    except asyncio.CancelledError:
-                        return
-                    except Exception:
-                        self.print_fail(traceback.format_exc())
-                    finally:
-                        command_done.reset(command_done_token)
-        finally:
-            current_stdout.reset(current_stdout_token)
-            current_monitor.reset(current_monitor_token)
-
-
-def auto_command_done(cmdfunc):
-    @functools.wraps(cmdfunc)
-    def _inner(ctx: click.Context, *args, **kwargs):
-        command_done_event = command_done.get()
-        try:
-            return cmdfunc(ctx, *args, **kwargs)
-        finally:
-            command_done_event.set()
-
-    return _inner
-
-
-def auto_async_command_done(cmdfunc):
-    @functools.wraps(cmdfunc)
-    async def _inner(ctx: click.Context, *args, **kwargs):
-        command_done_event = command_done.get()
-        try:
-            return await cmdfunc(ctx, *args, **kwargs)
-        finally:
-            command_done_event.set()
-
-    return _inner
-
-
-def custom_help_option(cmdfunc):
-    """
-    A custom help option to ensure setting `command_done_event`.
-    """
-
-    @auto_command_done
-    def show_help(ctx: click.Context, param: click.Parameter, value: bool) -> None:
-        if not value:
-            return
-        click.echo(ctx.get_help(), color=ctx.color)
-        ctx.exit()
-
-    return click.option(
-        "--help",
-        is_flag=True,
-        expose_value=False,
-        is_eager=True,
-        callback=show_help,
-        help="Show the help message",
-    )(cmdfunc)
-
-
-@monitor_cli.command(name="noop", hidden=True)
-@auto_command_done
-def do_noop(ctx: click.Context) -> None:
-    pass
-
-
-@monitor_cli.command(name="help", aliases=["?", "h"])
-@custom_help_option
-@auto_command_done
-def do_help(ctx: click.Context) -> None:
-    """Show the list of commands"""
-    click.echo(monitor_cli.get_help(ctx))
-
-
-@monitor_cli.command(name="signal")
-@click.argument("signame", type=str, shell_complete=complete_signal_names)
-@custom_help_option
-@auto_command_done
-def do_signal(ctx: click.Context, signame: str) -> None:
-    """Send a Unix signal"""
-    self: Monitor = ctx.obj
-    if hasattr(signal, signame):
-        os.kill(os.getpid(), getattr(signal, signame))
-        self.print_ok(f"Sent signal to {signame} PID {os.getpid()}")
-    else:
-        self.print_fail(f"Unknown signal {signame}")
-
-
-@monitor_cli.command(name="stacktrace", aliases=["st", "stack"])
-@custom_help_option
-@auto_command_done
-def do_stacktrace(ctx: click.Context) -> None:
-    """Print a stack trace from the event loop thread"""
-    self: Monitor = ctx.obj
-    stdout = _get_current_stdout()
-    tid = self._event_loop_thread_id
-    assert tid is not None
-    frame = sys._current_frames()[tid]
-    traceback.print_stack(frame, file=stdout)
-
-
-@monitor_cli.command(name="cancel", aliases=["ca"])
-@click.argument("taskid", shell_complete=complete_task_id)
-@custom_help_option
-@auto_async_command_done
-async def do_cancel(ctx: click.Context, taskid: str) -> None:
-    """Cancel an indicated task"""
-    self: Monitor = ctx.obj
-    try:
-        await self.cancel_monitored_task(taskid)
-        self.print_ok(f"Cancelled task {taskid}")
-    except ValueError as e:
-        self.print_fail(repr(e))
-
-
-@monitor_cli.command(name="exit", aliases=["q", "quit"])
-@custom_help_option
-@auto_command_done
-def do_exit(ctx: click.Context) -> None:
-    """Leave the monitor client session"""
-    raise asyncio.CancelledError("exit by user")
-
-
-@monitor_cli.command(name="console")
-@custom_help_option
-def do_console(ctx: click.Context) -> None:
-    """Switch to async Python REPL"""
-    self: Monitor = ctx.obj
-    if not self._console_enabled:
-        self.print_fail("Python console is disabled for this session!")
-        return
-
-    @auto_async_command_done
-    async def _console(ctx: click.Context) -> None:
-        h, p = self._host, self._console_port
-        log.info("Starting aioconsole at %s:%d", h, p)
-        app_session = get_app_session()
-        server = await console.start(
-            self._host,
-            self._console_port,
-            self.console_locals,
-            self._monitored_loop,
-        )
-        try:
-            await console.proxy(
-                app_session.input,
-                app_session.output,
-                self._host,
-                self._console_port,
-            )
-        except asyncio.CancelledError:
-            return
-        finally:
-            await console.close(server, self._monitored_loop)
-            log.info("Terminated aioconsole at %s:%d", h, p)
-            self.print_ok("The console session is closed.")
-
-    # Since we are already inside the UI's event loop,
-    # spawn the async command function as a new task and let it
-    # set `command_done_event` internally.
-    task = asyncio.create_task(_console(ctx))
-    self._console_tasks.add(task)
-
-
-@monitor_cli.command(name="ps", aliases=["p"])
-@click.option("-f", "--filter", "filter_", help="filter by coroutine or task name")
-@click.option("-p", "--persistent", is_flag=True, help="show only persistent tasks")
-@custom_help_option
-@auto_command_done
-def do_ps(
-    ctx: click.Context,
-    filter_: str,
-    persistent: bool,
-) -> None:
-    """Show task table"""
-    headers = (
-        "Task ID",
-        "State",
-        "Name",
-        "Coroutine",
-        "Created Location",
-        "Since",
-    )
-    self: Monitor = ctx.obj
-    stdout = _get_current_stdout()
-    table_data: List[Tuple[str, str, str, str, str, str]] = [headers]
-    tasks = self.format_live_task_list(filter_, persistent)
-    for task in tasks:
-        table_data.append(
-            (
-                task.task_id,
-                task.state,
-                task.name,
-                task.coro,
-                task.created_location,
-                task.since,
-            )
-        )
-    table = AsciiTable(table_data)
-    table.inner_row_border = False
-    table.inner_column_border = False
-    if filter_ or persistent:
-        stdout.write(
-            f"{len(tasks)} tasks running (showing {len(table_data) - 1} tasks)\n"
-        )
-    else:
-        stdout.write(f"{len(tasks)} tasks running\n")
-    stdout.write(table.table)
-    stdout.write("\n")
-    stdout.flush()
-
-
-@monitor_cli.command(name="ps-terminated", aliases=["pt", "pst"])
-@click.option("-f", "--filter", "filter_", help="filter by coroutine or task name")
-@click.option("-p", "--persistent", is_flag=True, help="show only persistent tasks")
-@custom_help_option
-@auto_command_done
-def do_ps_terminated(
-    ctx: click.Context,
-    filter_: str,
-    persistent: bool,
-) -> None:
-    """List recently terminated/cancelled tasks"""
-    headers = (
-        "Trace ID",
-        "Name",
-        "Coro",
-        "Since Started",
-        "Since Terminated",
-    )
-    self: Monitor = ctx.obj
-    stdout = _get_current_stdout()
-    table_data: List[Tuple[str, str, str, str, str]] = [headers]
-    tasks = self.format_terminated_task_list(filter_, persistent)
-    for task in tasks:
-        table_data.append(
-            (
-                task.task_id,
-                task.name,
-                task.coro,
-                task.started_since,
-                task.terminated_since,
-            )
-        )
-    table = AsciiTable(table_data)
-    table.inner_row_border = False
-    table.inner_column_border = False
-    if filter_ or persistent:
-        stdout.write(
-            f"{len(tasks)} tasks terminated " f"(showing {len(table_data) - 1} tasks)\n"
-        )
-    else:
-        stdout.write(f"{len(tasks)} tasks terminated (old ones may be stripped)\n")
-    stdout.write(table.table)
-    stdout.write("\n")
-    stdout.flush()
-
-
-@monitor_cli.command(name="where", aliases=["w"])
-@click.argument("taskid", shell_complete=complete_task_id)
-@custom_help_option
-@auto_command_done
-def do_where(ctx: click.Context, taskid: str) -> None:
-    """Show stack frames and the task creation chain of a task"""
-    self: Monitor = ctx.obj
-    stdout = _get_current_stdout()
-    depth = 0
-    task_id = int(taskid)
-    task = task_by_id(task_id, self._monitored_loop)
-    if task is None:
-        self.print_fail(f"No task {task_id}")
-        return
-    task_chain: List[asyncio.Task[Any]] = []
-    while task is not None:
-        task_chain.append(task)
-        task_ref = self._created_traceback_chains.get(task)
-        task = task_ref() if task_ref is not None else None
-    prev_task = None
-    for task in reversed(task_chain):
-        if depth == 0:
-            stdout.write(
-                "Stack of the root task or coroutine scheduled in the event loop (most recent call last):\n\n"
-            )
-        elif depth > 0:
-            assert prev_task is not None
-            stdout.write(
-                "Stack of %s when creating the next task (most recent call last):\n\n"
-                % _format_task(prev_task)
-            )
-        stack = self._created_tracebacks.get(task)
-        if stack is None:
-            stdout.write(
-                "  No stack available (maybe it is a native code, a synchronous callback function, or the event loop itself)\n"
-            )
-        else:
-            stack = _filter_stack(stack)
-            stdout.write("".join(traceback.format_list(stack)))
-        prev_task = task
-        depth += 1
-        stdout.write("\n")
-    task = task_chain[0]
-    stdout.write("Stack of %s (most recent call last):\n\n" % _format_task(task))
-    stack = _extract_stack_from_task(task)
-    if not stack:
-        stdout.write("  No stack available for %s" % _format_task(task))
-    else:
-        stdout.write("".join(traceback.format_list(stack)))
-    stdout.write("\n")
-
-
-@monitor_cli.command(name="where-terminated", aliases=["wt"])
-@click.argument("trace_id", shell_complete=complete_trace_id)
-@custom_help_option
-@auto_command_done
-def do_where_terminated(ctx: click.Context, trace_id: str) -> None:
-    """Show stack frames and the termination/cancellation chain of a task"""
-    self: Monitor = ctx.obj
-    stdout = _get_current_stdout()
-    depth = 0
-    tinfo_chain: List[TerminatedTaskInfo] = []
-    while trace_id is not None:
-        tinfo_chain.append(self._terminated_tasks[trace_id])
-        trace_id = self._canceller_chain.get(trace_id)  # type: ignore
-    prev_tinfo = None
-    for tinfo in reversed(tinfo_chain):
-        if depth == 0:
-            stdout.write(
-                "Stack of the root task or coroutine scheduled in the event loop (most recent call last):\n\n"
-            )
-        elif depth > 0:
-            assert prev_tinfo is not None
-            stdout.write(
-                "Stack of %s when creating the next task (most recent call last):\n\n"
-                % _format_terminated_task(prev_tinfo)
-            )
-        stack = tinfo.canceller_stack
-        if stack is None:
-            stdout.write(
-                "  No stack available (maybe it is a self-raised cancellation or exception)\n"
-            )
-        else:
-            stack = _filter_stack(stack)
-            stdout.write("".join(traceback.format_list(stack)))
-        prev_tinfo = tinfo
-        depth += 1
-        stdout.write("\n")
-    tinfo = tinfo_chain[0]
-    stdout.write(
-        "Stack of %s (most recent call last):\n\n" % _format_terminated_task(tinfo)
-    )
-    stack = tinfo.termination_stack
-    if not stack:
-        stdout.write(
-            "  No stack available for %s (the task has run to completion)"
-            % _format_terminated_task(tinfo)
-        )
-    else:
-        stdout.write("".join(traceback.format_list(stack)))
-    stdout.write("\n")
-
 
 def start_monitor(
     loop: asyncio.AbstractEventLoop,
     *,
     monitor_cls: Type[Monitor] = Monitor,
     host: str = MONITOR_HOST,
-    port: int = MONITOR_PORT,
+    port: int = MONITOR_TERMUI_PORT,  # kept the name for backward compatibility
     console_port: int = CONSOLE_PORT,
+    webui_port: int = MONITOR_WEBUI_PORT,
     console_enabled: bool = True,
     hook_task_factory: bool = False,
     max_termination_history: Optional[int] = None,
@@ -995,12 +619,15 @@ def start_monitor(
     m = monitor_cls(
         loop,
         host=host,
-        port=port,
+        termui_port=port,
+        webui_port=webui_port,
         console_port=console_port,
         console_enabled=console_enabled,
         hook_task_factory=hook_task_factory,
-        max_termination_history=max_termination_history
-        or get_default_args(monitor_cls.__init__)["max_termination_history"],
+        max_termination_history=(
+            max_termination_history
+            or get_default_args(monitor_cls.__init__)["max_termination_history"]
+        ),
         locals=locals,
     )
     m.start()
